Struct slabmalloc_safe::ZoneAllocator

pub struct ZoneAllocator {
    pub heap_id: usize,
    /* private fields */
}

A zone allocator for arbitrary sized allocations.

Has a bunch of SCAllocator and through that can serve allocation requests for many different object sizes up to (MAX_SIZE_CLASSES) by selecting the right SCAllocator for allocation and deallocation.

The allocator provides the refill function refill to provide the underlying SCAllocator with memory.

Fields
heap_id: usize
Implementations

impl ZoneAllocator

pub const MAX_ALLOC_SIZE: usize = 8_112usize

Maximum size that allocated within 2 pages. (8 KiB - metadata) This is also the maximum object size that this allocator can handle.

pub const MAX_BASE_ALLOC_SIZE: usize = 8_112usize

Maximum size which is allocated with ObjectPages8k (8 KiB).

e.g. this is 8 KiB - bytes of meta-data.

pub const MAX_BASE_SIZE_CLASSES: usize = 11usize

How many allocators of type SCAllocator we have.

pub const BASE_ALLOC_SIZES: [usize; 11] = _

The set of sizes the allocator has lists for.

pub fn new(heap_id: usize) -> ZoneAllocator

pub fn get_max_size(current_size: usize) -> Option<usize>

Return maximum size an object of size current_size can use.

Used to optimize realloc.

impl ZoneAllocator

pub fn retrieve_empty_page(
&mut self,
heap_empty_page_threshold: usize
) -> Option<MappedPages8k>

Returns a MappedPages8k from the SCAllocator with the maximum number of empty pages, if there are more empty pages than the threshold.

pub fn exchange_pages_within_heap(
&mut self,
layout: Layout
) -> Result<(), &'static str>

pub fn empty_pages(&self) -> usize

The total number of empty pages in this zone allocator

pub fn allocate(&mut self, layout: Layout) -> Result<NonNull<u8>, &'static str>

Allocate a pointer to a block of memory described by layout.

pub unsafe fn deallocate(
&mut self,
ptr: NonNull<u8>,
layout: Layout
) -> Result<(), &'static str>

Deallocates a pointer to a block of memory, which was previously allocated by allocate.

Arguments

ptr - Address of the memory location to free.

layout - Memory layout of the block pointed to by ptr.

Safety

The caller must ensure that ptr argument is returned from Self::allocate() and layout argument is correct.

pub fn refill(
&mut self,
layout: Layout,
mp: MappedPages8k
) -> Result<(), &'static str>

Refills the SCAllocator for a given Layout with a MappedPages8k.

Safety

MappedPages should not be in use by any other heap.